EVENTBUS + разделение по файлам + readme
