+++ public/index.html
@@ -4,7 +4,7 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>HotelScanner</title>
+    <title>HostelScan</title>
     <link rel="stylesheet" href="styles.css" />
 </head>
