+++ public/js/components/home/homeController.js
@@ -1,0 +1,10 @@
+import HomeView from './homeView'
+
+export default class HomeController {
+    constructor(parent) {
+        this._view = new HomeView(parent);
+    }
+    activate() {
+        this._view.render()
+    }
+}
